implement token rotation
implements a system of token rotation to help secure user authentication in the application. it provides a user with an access_token and a refresh_token, as well as a way to refresh and validate the access_token. the lifetime of the access_token is 1 hour and the lifetime of the refresh_token is 1 day.

modifies the `readme.md` file to update the task list, changing the status of the token rotation task from `[ ]` to `[x]` to indicate that it has been completed.