refactor authentication and events app
refactors the authentication and events app. the authentication app url is changed to `auth/` and the events app is renamed from `pages` to `events`.

the `authentication/urls.py` is changed to have `auth/` in the path. the `core/settings.py` is changed to include `events` in the `installed_apps` list, and the `core/urls.py` is changed to have `api/` in the path for the authentication app. the `pages/` directory is renamed to `events/` and the relevant files are updated accordingly. the `pages/__pycache__/` directory is also deleted.