add events url to core urls
adds the events urls to core urls by adding `path('api/', include('events.urls'))` to the `urlpatterns` list in `core/urls.py`, and a new `events.urls` module with a single `events` view.

allows for the `events` view to be accessed from the `api/events` endpoint, which returns a `status: ok` response.