add a __str__ method to event model
adds the __str__ method to the event model in the events/models.py file. this new method returns the event's name when it is called and it will be used for the event representation in other parts of the application.

the modified section of the file is as follows:
```diff
@@ -21,3 +21,6 @@ class event(models.model):
         related_name='subscriptions',
         blank=true
     )
+
+    def __str__(self):
+        return self.name
```
the new __str__ method will be used to represent the event in other areas of the application.