add ability to create events
adds the ability to create events in the app's database. this includes adding a new `eventcreateserializer` and `create` method in the `eventsviewset` to handle the request and response. the `created_by` field is set to the requesting user.

the diff includes changes to the `db.sqlite3` binary file, `events/serializers.py` and `events/views.py` files and the `readme.md` file. the `serializers.py` file has a new `eventcreateserializer` class and the `views.py` file has a new `create` method that sets the `created_by` field to the requesting user. the `readme.md` file updates the status of the feature.