retrieve method to eventsviewset
adds a `retrieve` method to the `eventsviewset` class in order to retrieve a single event. the method will first get the event and serialize it, then check if the request user is the creator of the event; if not, the user will receive an unauthorized response. otherwise, the request user will receive an ok response with the event's data.

```diff
diff --git a/events/views.py b/events/views.py
index cb503495c9cd..dc9f5516db48 100644
--- a/events/views.py
+++ b/events/views.py
@@ -10,6 +10,21 @@ class eventsviewset(viewsets.modelviewset):
     serializer_class = eventserializer
     permission_classes = (custompermissions,)

+    def retrieve(self, request, pk=none):
+        event = event.objects.get(id=pk)
+        serialized = eventserializer(event)
+        if event.created_by != request.user:
+            return response(
+                status=status.http_401_unauthorized,
+                data={
+                    'message': 'you are not allowed to see this event.'
+                }
+            )
+        return response(
+            status=status.http_200_ok,
+            data=serialized.data
+        )
+
     def create(self, request, *args, **kwargs):
         data = copy(self.request.data)
         data['created_by'] = self.request.user.id
```