removing event retrieve method
removes the `retrieve` method from the `eventsviewset` class in `views.py`. this change will remove the ability for users to retrieve a specific event unless they are the creator of that event.

```diff
diff --git a/events/views.py b/events/views.py
index dc9f5516db48..cb503495c9cd 100644
--- a/events/views.py
+++ b/events/views.py
@@ -10,21 +10,6 @@ class eventsviewset(viewsets.modelviewset):
     serializer_class = eventserializer
     permission_classes = (custompermissions,)

-    def retrieve(self, request, pk=none):
-        event = event.objects.get(id=pk)
-        serialized = eventserializer(event)
-        if event.created_by != request.user:
-            return response(
-                status=status.http_401_unauthorized,
-                data={
-                    'message': 'you are not allowed to see this event.'
-                }
-            )
-        return response(
-            status=status.http_200_ok,
-            data=serialized.data
-        )
-
     def create(self, request, *args, **kwargs):
         data = copy(self.request.data)
         data['created_by'] = self.request.user.id
```

this commit removes the `retrieve` method from the `eventsviewset` class in `views.py`. this change will remove the ability for users to retrieve a specific event unless they are the creator of that event. it also adds a check to the `create` method to ensure that the `created_by` field is populated with the user creating the event.