add ordering to eventsviewset and implement subscribed action
adds ordering to the eventsviewset and implements the subscribed action. the ordering is added to the queryset for the eventsviewset such that it is ordered by the id field. additionally, the subscribed action is implemented which returns all the subscriptions for the current user. this action is accessible via the get method and is a detail=false.