changed event end date default
changes the default value of the event's end date from null to two days after the current date. this will ensure that events are scheduled for at least two days in the future.

by default, the event object's end date was set to null, meaning the user had to fill it in manually. now it is set to two days after the current date, ensuring that events are scheduled for a minimum of two days. this makes it easier and faster to create events, as the user will not have to manually enter a new end date.

this commit makes the following changes to the `events/models.py` file:
```diff
diff --git a/events/models.py b/events/models.py
index de9ef185fb05..c1314866cdc0 100644
--- a/events/models.py
+++ b/events/models.py
@@ -7,7 +7,7 @@ class event(models.model):
     description = models.textfield()
     status = models.charfield(max_length=100)
     start_date = models.datetimefield(default=datetime.now() + timedelta(days=1))
-    end_date = models.datetimefield()
+    end_date = models.datetimefield(default=datetime.now() + timedelta(days=2))
     location = models.charfield(max_length=100)
     capacity = models.integerfield(default=20)
     is_public = models.booleanfield(default=true)
```