add unsubscribe action to eventsviewset
adds a new action to the eventsviewset to enable users to unsubscribe from an event. the new action takes a request and a primary key, checks if the user is already subscribed to the event and then removes them from the list of subscribers if they are. finally, a response is sent back to the user confirming their success or failure.

the database file `db.sqlite3` was updated to reflect this change, as well as the `events/views.py` file. the `events/views.py` file was modified with the following changes:
- added an `@action` with a `methods` parameter set to `post`, and a `detail` parameter set to `true`
- added a `unsubscribe` method that takes a request and primary key as parameters
- within the `unsubscribe` method, checks if the user is already subscribed to the event and removes them from the list of subscribers if they are
- added a response confirming the success or failure of the operation