add feature to register and unregister from events
adds the feature where users can register and unregister from events. this is a feature that allows users to register and unregister from events they have created as long as the event is in the future. this feature adds more flexibility for users when managing their events.

the following changes were made in this commit:

- updated readme file to add a checkbox saying that users can register and unregister from events
- added the feature to the codebase which allows users to register and unregister from events