updating authentication app response values
updates the authentication app to include a custom token obtain pair serializer. it also includes changes to the authentication urls, serializers, views, and settings file. the changes include the addition of a custom token obtain pair serializer, which adds the username, email, first name, and last name of the user to the returned token data. it also includes changes to the authentication urls, serializers, views, and settings file to enable cross-origin resource sharing (cors).