add new users to the database
adds three new users to the database. the users are alexsander falcucci, aurora gaeta, and christopher waltz.

the changes were made to the db.sqlite3 file, and a new file - fixtures/users.json - was created. this commit also changes the start and end date of the events in the event model, using a migration file (0003_alter_event_end_date_alter_event_start_date.py). the new dates are set to may 12th and 13th, 2023.