update database and documentation
updates the database and documentation for the [event-manager-api](https://github.com/falcucci/event-manager-api) project. the database was updated to ensure that all events are properly tracked and stored. additionally, the documentation was updated to include a new section on features, as well as more detailed instructions on how to use the api. finally, some minor changes were made to the to-do list to reflect the current status of the project.