add swagger documentation to project
adds the swagger documentation to the project. this includes adding the [drf_yasg](https://drf-yasg.readthedocs.io/en/stable/) package to `installed_apps` in `settings.py`, configuring the schema view in `urls.py` and adding an openapi info object in `urls.py` to provide a description of the api and contact information.

the changes made in this commit are:

* add `drf_yasg` to `installed_apps` in `settings.py`
* configure schema view in `urls.py`
* add openapi info object in `urls.py` to provide a description of the api and contact information