add installation and running of django project
adds an installation and running guide for a django project. it includes instructions on cloning the project, creating a virtual environment, installing dependencies, migrating the database and running the project. it also includes instructions on loading fixtures.

overall, this commit provides detailed instructions on setting up and running a django project.