refactor event capacity verification
adds the necessary logic to check that an event's capacity has not been exceeded before allowing a user to subscribe to it. specifically, it adds an extra check to the `eventsviewset` class in the `views.py` file, which will return a `400 bad request` status code if the event is full.

```diff
diff --git a/events/views.py b/events/views.py
index dba4eaf703dc..5eb729751fce 100644
--- a/events/views.py
+++ b/events/views.py
@@ -75,6 +75,12 @@ class eventsviewset(viewsets.modelviewset):
                 data={"detail": "event has already started."}
             )

+        if event.capacity == event.subscribers.count():
+            return response(
+                status=status.http_400_bad_request,
+                data={"detail": "event is full."}
+            )
+
         event.subscribers.add(request.user)
         return response(
             status=status.http_200_ok,
```