adding logic to manage event capacity
adds the necessary logic to manage an event capacity. if the event reaches the maximum number of registered attendees, an error will be returned to a user trying to register.

by adding this logic, the api will be able to better manage events and ensure that the maximum number of attendees is not exceeded. this will prevent any issues with overcrowding or other problems that could arise if too many people are attending an event.