add .vercel to .gitignore and fix db config
adds a new line to the .gitignore file, as well as updates the core/settings.py and core/wsgi.py files and adds a .env file. it also updates the requirements.txt file.

adds the new line `.vercel` to the .gitignore file, updates the `wsgi_application` variable in `core/settings.py` and `core/wsgi.py` to `'core.wsgi.app'` and adds a `.env` file to the project. it also updates the requirements.txt file to include `coverage`, `dj-database-url`, `psycopg2-binary`, `python-decouple` and `typing-extensions`.