update docs todo list
adds the tests that were missing from the project. tests are an important part of any software development project as they help ensure that any changes made to the codebase do not cause unexpected behavior.

the codebase was updated with the following changes:

- [x] tests

these tests provide coverage for the new logic that was added to manage an event capacity. they also provide coverage for the different filtering endpoints. finally, they provide coverage for the new frontend added to consume the api.